6/inventory_backend/app/db.py
import psycopg2
from psycopg2.extras import RealDictCursor
import os

DB_CONFIG = {
    "dbname": os.getenv("POSTGRES_DB", "books_db"),
    "user": os.getenv("POSTGRES_USER", "postgres"),
    "password": os.getenv("POSTGRES_PASSWORD", "books123"),
    "host": os.getenv("POSTGRES_HOST", "localhost"),
    "port": os.getenv("POSTGRES_PORT", 5434)
}

def get_connection():
    conn = psycopg2.connect(**DB_CONFIG, cursor_factory=RealDictCursor)
    return conn
