6/inventory_backend/requirements.txt
fastapi[standard]
uvicorn[standard]
psycopg2-binary
python-dotenv
httpx
pydantic